type: Added a game title and display this title in an opening modal at start of game: Don't Bug Me!
Modified the application code (app.js) to display the game title at start of the game
using a modal. Also modified index.html to include the modal with title and modified
the html title field. Added a little bit of style to the title being displayed in the
modal, in the css/style.css file. In the README, simply added extra # to make the subheadings
smaller.